feat: retun a stream object from stream_audio
Previously you had to supply an async buffer to "stream" audio, but that
proved to be challenging when attempting to play the audio in real time.
We have now added a stream_audio method that returns a stream and
renamed the original method to write_audio_stream so it's clearer what
it actually does.

We've done a serious cleanup of README and examples and made sure all
the examples are compiled during CI runs.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>